Reload now reloads commands too
