+++ commands/reload.js
@@ -1,11 +1,22 @@
+const fs = require('fs');
 const Discord = require('discord.js');
 
 module.exports =  {
-    name: "loademotes",
-    alias: ["reload"],
+    name: "reload",
+    alias: [],
     execute(client) {
+        client.commands = new Discord.Collection();
         client.emotes = new Discord.Collection();
         client.comboemotes = new Discord.Collection();
+
+        const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
+        for (file of commandFiles) {
+            const command = require(`./${file}`);
+            client.commands.set(command.name, command);
+            for (alias of command.alias) {
+                client.commands.set(alias, command);
+            }
+        }
         
         for (guild of client.guilds.values()) {
             for (emoji of guild.emojis.values()) {
